k8s/src/satellite-interface/requirements.txt
python-dotenv==1.2.1
pika==1.3.2
mysql-connector-python==9.5.0
opentelemetry-api==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-exporter-otlp==1.39.1
opentelemetry-instrumentation==0.60b1
opentelemetry-instrumentation-requests==0.60b1
opentelemetry-instrumentation-pika==0.60b1
confluent-kafka==2.12.2